requirements.txt
opentelemetry-distro
flask
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
prometheus-flask-exporter
